+++ fixtures/themed/src/index.ts
@@ -115,11 +115,3 @@ function render() {
 }
 
 render();
-
-// @ts-expect-error
-if (module.hot) {
-  // @ts-expect-error
-  module.hot.accept(['./shared.css', './styles.css', './themes.css'], () => {
-    render();
-  });
-}