+++ test-helpers/src/startFixture/types.ts
@@ -2,10 +2,12 @@ export type BuildType =
   | 'browser'
   | 'mini-css-extract'
   | 'style-loader'
-  | 'esbuild';
+  | 'esbuild'
+  | 'vite';
 
 export interface TestServer {
   type: BuildType;
   url: string;
   close: () => Promise<void>;
+  stylesheet?: string;
 }