+++ tests/sprinkles/index.css.ts
@@ -50,7 +50,7 @@ export const conditionalAtomicStyles = createAtomicStyles({
     display: ['block', 'none', 'flex'],
     paddingTop: spacing,
     paddingBottom: spacing,
-    opacity: [0, 1],
+    opacity: [0, 1] as const,
   },
   shorthands: {
     paddingY: ['paddingBottom', 'paddingTop'],